from reduce function to basic redux
